EQUAL working fine

+++ lex.yy.c
@@ -802,7 +802,7 @@ YY_RULE_SETUP
 case 2:
 YY_RULE_SETUP
 #line 10 "compiler.l"
-{ yylval.num = atoi(yytext); return NUM; }
+{ yylval.num = atoll(yytext); return NUM; }
 	YY_BREAK
 case 3:
 YY_RULE_SETUP
